training/requirements.txt
torch>=2.0
transformers>=4.36
peft>=0.7
trl>=0.7
datasets>=2.16
accelerate>=0.25
bitsandbytes>=0.41
pyyaml>=6.0
